% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_HR_GAM.R
\name{fit_HR_GAM}
\alias{fit_HR_GAM}
\title{Title}
\usage{
fit_HR_GAM(
  d,
  Y,
  X,
  age,
  W = NULL,
  forcedW = NULL,
  V = NULL,
  id = "clusterid",
  pval = 0.2,
  print = TRUE
)
}
\arguments{
\item{forcedW}{Set to NA to skip forcing any variables, don't include to skip prescreening of "tr" and variables starting with age_, ageday_, or childage_. Or manually set with a vector of variable names.}

\item{print}{}
}
\value{

}
\description{
Title
}
