% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_RE_gam.R
\name{fit_RE_gam}
\alias{fit_RE_gam}
\title{Title}
\usage{
fit_RE_gam(
  d,
  Y,
  X,
  W = NULL,
  V = NULL,
  id = "clusterid",
  family = "gaussian",
  pval = 0.2,
  print = TRUE
)
}
\arguments{
\item{print}{}
}
\value{

}
\description{
Title
}
