% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sig_heatmap.R
\name{plot_sig_heatmap}
\alias{plot_sig_heatmap}
\title{Title}
\usage{
plot_sig_heatmap(
  d,
  pval_var = "Pval",
  title = "",
  Outcome = "Outcome",
  Exposure = "Exposure",
  print.est = T,
  print.ci = F,
  null = 0
)
}
\arguments{
\item{print.ci}{}
}
\description{
Title
}
