% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gam_HR.R
\name{predict_gam_HR}
\alias{predict_gam_HR}
\title{Title}
\usage{
predict_gam_HR(
  fit,
  d,
  quantile_diff = c(0.25, 0.75),
  Xvar,
  Yvar,
  binaryX = FALSE
)
}
\arguments{
\item{binaryX}{}
}
\description{
Title
}
